implemented crud

+++ ContosoU/Controllers/StudentController.cs
@@ -44,6 +44,17 @@ namespace ContosoU.Controllers
             ViewData["EmailSortParm"] = sortOrder == "email" ? "email_desc" : "email";
             ViewData["DateSortParm"] = sortOrder == "date" ? "date_desc" : "date";
 
+            //same as above
+            //if(sortOrder == "fname")
+            //{
+            //    sortOrder = "fname_desc";
+            //}
+            //else
+            //{
+            //    sortOrder = "fname";
+            //}
+            //ViewData["FNameSortParm"] = sortOrder;
+
             //part 2: Filtering
             if(searchString == null)
             {
@@ -111,11 +122,37 @@ namespace ContosoU.Controllers
         {
             if (id == null)
             {
+                /* status codes:
+                 * Success: return ok() <- HTTP status code 200
+                 *          return Created() <- HTTP status code 201
+                 *          return NoContent() <- HTTP status code 204
+                 * 
+                 * Client Error: return BadRequest() <- HTTP status 400
+                 *               return Unauthorized() <- HTTP status 401
+                 *               return NotFound() <- HTTP status 404
+                 * 
+                 */
                 return NotFound();
             }
-
+            //fcomeau: update to include related date(enrollement)
+            //var student = await _context.Students
+            //    .SingleOrDefaultAsync(m => m.ID == id);
             var student = await _context.Students
+                .Include(s => s.Enrollments).ThenInclude(c => c.Course)
+                .AsNoTracking()
                 .SingleOrDefaultAsync(m => m.ID == id);
+            /*
+            * ============================================================== No-tracking queries ======================================
+            * When a database context retrieves table rows and creates entity objects that represent them, by default it keeps track
+            * of whether 
+            * the entities in memory are in sync with what's in the database. The data in memory acts as a cache and is used when you
+            * update an entity. 
+            * This caching is often unnecessary in a web application because context instances are typically short-lived 
+            * (a new one is created and disposed for each request) and the context that reads an entity is typically disposed before 
+            * that entity is used again.
+            * 
+            * Ref:  https://docs.microsoft.com/en-us/ef/core/querying/tracking
+            */
             if (student == null)
             {
                 return NotFound();
@@ -135,14 +172,26 @@ namespace ContosoU.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("EnrollmentDate,ID,LastName,FirstName,Email")] Student student)
+        //fcomeau: remove the ID from the bind : ID is the pk as well as IDENTITY
+        public async Task<IActionResult> Create([Bind("EnrollmentDate,LastName,FirstName,Email")] Student student)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(student);
-                await _context.SaveChangesAsync();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    _context.Add(student);
+                     await _context.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
             }
+            catch (Exception)
+            {
+
+                //log the error (by uncommenting the ex variable) and write to a log file.
+                //return a ModelStateError back to user
+                ModelState.AddModelError("", "Unable to save changes. " + "Please try again");
+            }
+            
             return View(student);
         }
 
@@ -165,41 +214,47 @@ namespace ContosoU.Controllers
         // POST: Student/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("EnrollmentDate,ID,LastName,FirstName,Email")] Student student)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != student.ID)
+            if (id == null)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            //find student to be updated(Bind)
+            var studentToUpdate = await _context.Students.SingleOrDefaultAsync(s => s.ID == id);
+            //try to update this student
+            if (await TryUpdateModelAsync<Student>(studentToUpdate, "", s => s.FirstName, s => s.LastName, s => s.Email,
+                s => s.EnrollmentDate))
             {
                 try
                 {
-                    _context.Update(student);
-                    await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync();//save changes to database
+                    return RedirectToAction("Index");//redirect user back to index route
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException)
                 {
-                    if (!StudentExists(student.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //log the error (by uncommenting the ex variable) and write to a log file.
+                    //return a ModelStateError back to user
+                    ModelState.AddModelError("", "Unable to save changes. " + "Please try again");
+
                 }
-                return RedirectToAction("Index");
             }
-            return View(student);
-        }
+
+            //return the view and attach the studentToUpdate model
+            return View(studentToUpdate);
+        }  
 
         // GET: Student/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? savechangeserror = false )
         {
+            /*this code accepts an optional boolean parameter that indicates whether the method was called after a delete
+             * failure(failure saving changes back to database
+             * when its called by the HTTPPost Delete Method in responce to a database update error, this parameter will
+             * be passed in set to true
+             */
             if (id == null)
             {
                 return NotFound();
@@ -211,7 +266,11 @@ namespace ContosoU.Controllers
             {
                 return NotFound();
             }
-
+            //fcomeau: return update error if necessary
+            if(savechangeserror.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] = "Delete failed! Try again later.";
+            }
             return View(student);
         }
 
@@ -220,10 +279,26 @@ namespace ContosoU.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var student = await _context.Students.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction("Index");
+            var student = await _context.Students.AsNoTracking().SingleOrDefaultAsync(m => m.ID == id);
+
+            //fcomeau: check if student exists
+            if(student == null)
+            {
+                return RedirectToAction("Index");
+            }
+
+            try
+            {
+                _context.Students.Remove(student);
+                await _context.SaveChangesAsync();
+                return RedirectToAction("Index");
+            }
+            catch (DbUpdateException)
+            {
+                //return user to delete get method passing the current student(ID) and a flag argument set to true for
+                //representing and save error record
+                return RedirectToAction("Delete", new { id = id, savechangeserror = true });
+            }
         }
 
         private bool StudentExists(int id)
